pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qa-uscomparer"
version = "0.1.0"
description = "Compare two Jira ticket definitions side-by-side using the Atlassian MCP server"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "e.videoqa035" }]
keywords = ["jira", "atlassian", "mcp", "qa", "comparison", "diff"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "mcp>=1.5.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "anyio>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[project.scripts]
qa-uscomparer = "qa_uscomparer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/qa_uscomparer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
